api/requirements.txt
aniso8601==8.0.0
attrs==20.2.0
autopep8==1.5.4
cffi==1.14.3
click==7.1.2
cryptography==3.1.1
Flask==1.1.2
Flask-Cors==3.0.9
Flask-RESTful==0.3.8
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.1
pyparsing==2.4.7
pytest==6.1.1
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
Werkzeug==1.0.1
